requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.0
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.21
pymongo==3.12.1
python-dotenv==0.19.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
yapf==0.31.0
